requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
supabase>=2.0.0

# Alternative data sources
polygon-api-client>=1.12.0
alpha-vantage>=2.3.1
quandl>=3.7.0

# Web framework
streamlit>=1.28.0
altair>=5.0.0
plotly>=5.0.0

# Email and notifications
# smtplib is part of Python standard library - no need to install

# Data processing and analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
pytz>=2023.3
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0